Move NormalizedJSON function to the only place it is used.
