+++ core/gethwrappers/generation/generate_link/utils/json_normalization.go
@@ -1,4 +1,4 @@
-package util
+package utils
 
 import (
 	"bufio"
